requirements.txt
python-dotenv
slack-bolt
openai